Update 04 build, removing broken RPM package management.
Symbiflow wrapped the commerical fork of Red Hat's old Anaconda RPM
package management system, and then symbiflow's maintainer made it a subrepo
of skywater-pdk that gets called check build prerequisites... except it's
checking rpm on dpkg based system and not finding stuff that's there.
(And it's checking at build time, not at ./configure time.)

The fix: delete the subrepo. The Makefile -include trying to use it drops
out if the file isn't there. We already provided the build prerequisites.
To do this, we need to check out the skywater-pdk repo, modify it, and
specify the replacement repo when calling the open_pdk build. Note that
we need to remove it twice: it's a subrepo and the Makefile is trying to
check it out.